+++ src/main/java/com/zen/fogman/ManFromTheFogClient.java
@@ -1,10 +1,12 @@
 package com.zen.fogman;
 
+import com.zen.fogman.client.ManClientTick;
 import com.zen.fogman.client.StaticOverlayHud;
 import com.zen.fogman.entity.ModRenderers;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 
 @Environment(EnvType.CLIENT)
@@ -15,5 +17,6 @@ public class ManFromTheFogClient implements ClientModInitializer {
 		ModRenderers.registerRenderers();
 
 		HudRenderCallback.EVENT.register(new StaticOverlayHud());
+		ClientTickEvents.END_CLIENT_TICK.register(new ManClientTick());
 	}
 }