+++ src/main/java/com/zen/fogman/entity/custom/TheManEntityHallucination.java
@@ -1,6 +1,9 @@
 package com.zen.fogman.entity.custom;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.server.world.ServerWorld;
@@ -15,12 +18,12 @@ public class TheManEntityHallucination extends TheManEntity {
     }
 
     @Override
-    public void addEffectsToClosePlayers(ServerWorld world, Vec3d pos, @Nullable Entity entity, int range) {
+    public void clientTick(MinecraftClient client) {
 
     }
 
     @Override
-    public void playChaseSound(MinecraftClient client) {
+    public void addEffectsToClosePlayers(ServerWorld world, Vec3d pos, @Nullable Entity entity, int range) {
 
     }