+++ src/main/java/com/zen/fogman/item/custom/TearOfTheManItem.java
@@ -1,10 +1,7 @@
 package com.zen.fogman.item.custom;
 
-import com.zen.fogman.ManFromTheFog;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.damage.DamageSource;
-import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.item.Item;